Temporarily removing super as allowed keyword in property name as it causes some issues when used after a dot.
For instance, currently Groovy doesn't like the following:

def m = [super: "cool"] // fine
assert m.super == "cool" // not fine :-(

2 (misleading) compilation errors:

Inner classes referencing outer classes using super is not supported yet.
 at line: 2, column: 8

The usage of '.this' or '.super' requires an explicit class in front.
 at line: 2, column: 8

Would be good to find a way to just allow super as map key in that situation.

git-svn-id: http://svn.codehaus.org/groovy/trunk/groovy/groovy-core@20511 a5544e8c-8a19-0410-ba12-f9af4593a198